Change default layout behaviour
